getCollectionIdPath consolidation
