+++ frontend/src/metabase/common/components/Pickers/QuestionPicker/utils.ts
@@ -1,6 +1,5 @@
 import _ from "underscore";
 
-import { PERSONAL_COLLECTIONS } from "metabase/entities/collections/constants";
 import { isNullOrUndefined } from "metabase/lib/types";
 import type {
   Card,
@@ -15,46 +14,6 @@ import type {
   QuestionPickerValue,
   QuestionPickerValueModel,
 } from "./types";
-
-export const getCollectionIdPath = (
-  collection: Pick<
-    QuestionPickerItem,
-    "id" | "location" | "is_personal" | "effective_location" | "model"
-  >,
-  userPersonalCollectionId?: CollectionId,
-): CollectionId[] => {
-  if (collection.id === "root" || collection.id === null) {
-    return ["root"];
-  }
-
-  if (collection.id === PERSONAL_COLLECTIONS.id) {
-    return ["personal"];
-  }
-
-  if (typeof collection.id === "string") {
-    console.error("Invalid collection id", collection.id);
-    return [];
-  }
-
-  const location = collection?.effective_location ?? collection?.location;
-  const pathFromRoot: CollectionId[] =
-    location?.split("/").filter(Boolean).map(Number) ?? [];
-
-  const isInUserPersonalCollection =
-    userPersonalCollectionId &&
-    (collection.id === userPersonalCollectionId ||
-      pathFromRoot.includes(userPersonalCollectionId));
-
-  const id = collection.model === "collection" ? collection.id : -collection.id;
-
-  if (isInUserPersonalCollection) {
-    return [...pathFromRoot, id];
-  } else if (collection.is_personal) {
-    return ["personal", ...pathFromRoot, id];
-  } else {
-    return ["root", ...pathFromRoot, id];
-  }
-};
 
 export const getStateFromIdPath = ({
   idPath,
